+++ ExampleSwift/ExampleSwift/Controllers/ViewController.swift
@@ -6,7 +6,6 @@
 //  Copyright © 2018 Juan Sebastian Sanzone. All rights reserved.
 // Check full documentation: http://mercadopago.github.io/px-ios/v4/
 //
-
 import UIKit
 
 #if PX_PRIVATE_POD
@@ -19,19 +18,9 @@ import UIKit
 class ViewController: UIViewController {
     private var checkout: MercadoPagoCheckout?
     
-    // Collector Public Key
-    private let publicKey : String = "TEST-e28d5a35-dece-45c9-9618-e8cc5dec6c42"
-    
-    // Preference ID
-    private let preferenceId : String = "656525290-2184df26-5283-41da-9a1c-9346735e73d2" //"656525290-7bda964b-26d9-4352-a04c-1b04801627ee"
-    
-    // Payer private key
-    private let privateKey : String = "TEST-7215456036166479-101915-223c2540d9126044887a840cfae544a9-660760647"
-    
     @IBAction func initDefault(_ sender: Any) {
         // runMercadoPagoCheckout()
-//         runMercadoPagoCheckoutWithLifecycle()
-        runMercadoPagoCheckoutWithLifecycleAndCustomProcessor()
+        runMercadoPagoCheckoutWithLifecycle()
     }
 
     override func viewDidLoad() {
@@ -47,7 +36,7 @@ class ViewController: UIViewController {
 
     private func runMercadoPagoCheckout() {
         // 1) Create Builder with your publicKey and preferenceId.
-        let builder = MercadoPagoCheckoutBuilder(publicKey: publicKey, preferenceId: preferenceId).setLanguage("es")
+        let builder = MercadoPagoCheckoutBuilder(publicKey: "TEST-4763b824-93d7-4ca2-a7f7-93539c3ee5bd", preferenceId: "243966003-0812580b-6082-4104-9bce-1a4c48a5bc44").setLanguage("es")
 
         // 2) Create Checkout reference
         checkout = MercadoPagoCheckout(builder: builder)
@@ -57,86 +46,15 @@ class ViewController: UIViewController {
             checkout?.start(navigationController: myNavigationController)
         }
     }
-    
-    private func runMercadoPagoCheckoutWithLifecycleAndCustomProcessor() {
-        
-        // Create charge rules
-        var pxPaymentTypeChargeRules : [PXPaymentTypeChargeRule] = []
-        
-        pxPaymentTypeChargeRules.append(PXPaymentTypeChargeRule.init(paymentTypeId: PXPaymentTypes.ACCOUNT_MONEY.rawValue, amountCharge: 10.00 ))
-
-//        // Charge rule with custom dialog
-//
-//        let alertController : UIAlertController = UIAlertController.init(title: "Detalle del cargo", message: "Este es el detalle del cargo que visualizás", preferredStyle: .alert)
-//
-//        pxPaymentTypeChargeRules.append(PXPaymentTypeChargeRule.init(paymentTypeId: PXPaymentTypes.CREDIT_CARD.rawValue, amountCharge: 15.00, detailModal: alertController ))
-//
-  
-        // Free charge rule
-        pxPaymentTypeChargeRules.append(PXPaymentTypeChargeRule.init(paymentTypeId: PXPaymentTypes.CREDIT_CARD.rawValue, message: "Mensaje de resaltado"))
-        
-        // Create an instance of your custom payment processor
-        let paymentProcessor : PXPaymentProcessor = CustomPaymentProcessor()
-        
-        // Create a payment configuration instance using the recently created payment processor
-        let paymentConfiguration = PXPaymentConfiguration(paymentProcessor: paymentProcessor)
-        
-        // Add charge rules
-        paymentConfiguration.addChargeRules(charges: pxPaymentTypeChargeRules)
-        
-        // Create a Builder with your publicKey, preferenceId and paymentConfiguration
-        let builder = MercadoPagoCheckoutBuilder(publicKey: publicKey, preferenceId: preferenceId, paymentConfiguration: paymentConfiguration).setLanguage("es")
-        
-//        let checkoutPreference = PXCheckoutPreference.init(preferenceId: preferenceId)
-//
-//        let builder = MercadoPagoCheckoutBuilder.init(publicKey: publicKey, checkoutPreference: checkoutPreference, paymentConfiguration: paymentConfiguration)
-        
-        // Instantiate a configuration object
-        let configuration = PXAdvancedConfiguration()
-         
-        // Set expressEnabled true to use one-tap instead of groups flow
-        configuration.expressEnabled = true
-        
-        // Add custom PXDynamicViewController component
-        configuration.dynamicViewControllersConfiguration = [CustomPXDynamicComponent()]
-        
-        // Configure the builder object
-        builder.setAdvancedConfiguration(config: configuration)
-        
-        // Set the payer private key
-        builder.setPrivateKey(key: privateKey)
-
-        // Create Checkout reference
-        checkout = MercadoPagoCheckout(builder: builder)
-
-        // Start with your navigation controller.
-        if let myNavigationController = navigationController {
-            checkout?.start(navigationController: myNavigationController, lifeCycleProtocol: self)
-        }
-    }
 
     private func runMercadoPagoCheckoutWithLifecycle() {
-        // Create Builder with your publicKey and preferenceId.
-        let builder = MercadoPagoCheckoutBuilder(publicKey: publicKey, preferenceId: preferenceId).setLanguage("es")
-        
-        let configuration = PXAdvancedConfiguration()
-        
-//        configuration.expressEnabled = true
-        
-        builder.setAdvancedConfiguration(config: configuration)
-        
-        // Set the payer private key
-        builder.setPrivateKey(key: privateKey)
-        
-        // Add custom translations (px_custom_texts)
-        builder.addCustomTranslation(.pay_button, withTranslation: "Pagar custom")
-        builder.addCustomTranslation(.pay_button_progress, withTranslation: "Pagando custom...")
-        builder.addCustomTranslation(.total_to_pay_onetap, withTranslation: "Total a pagar custom")
+        // 1) Create Builder with your publicKey and preferenceId.
+        let builder = MercadoPagoCheckoutBuilder(publicKey: "TEST-4763b824-93d7-4ca2-a7f7-93539c3ee5bd", preferenceId: "243966003-0812580b-6082-4104-9bce-1a4c48a5bc44").setLanguage("es")
 
-        // Create Checkout reference
+        // 2) Create Checkout reference
         checkout = MercadoPagoCheckout(builder: builder)
 
-        // Start with your navigation controller.
+        // 3) Start with your navigation controller.
         if let myNavigationController = navigationController {
             checkout?.start(navigationController: myNavigationController, lifeCycleProtocol: self)
         }