+++ MercadoPagoSDK/MercadoPagoSDK/Services/Models/PXPointsAndDiscounts.swift
@@ -17,8 +17,13 @@ struct PXPointsAndDiscounts: Decodable {
     let customOrder: Bool?
     let expenseSplit: PXExpenseSplit?
     let paymentMethodsImages: [String: String]?
+    let primaryButton: PXButton?
+    let secondaryButton: PXButton?
+    let redirectUrl: String?
+    let backUrl: String?
+    let autoReturn: PXAutoReturn?
 
-    init(points: PXPoints?, discounts: PXDiscounts?, crossSelling: [PXCrossSellingItem]?, viewReceiptAction: PXRemoteAction?, topTextBox: PXText?, customOrder: Bool?, expenseSplit: PXExpenseSplit?, paymentMethodsImages: [String: String]?) {
+    init(points: PXPoints?, discounts: PXDiscounts?, crossSelling: [PXCrossSellingItem]?, viewReceiptAction: PXRemoteAction?, topTextBox: PXText?, customOrder: Bool?, expenseSplit: PXExpenseSplit?, paymentMethodsImages: [String: String]?, primaryButton: PXButton?, secondaryButton: PXButton?, redirectUrl: String?, backUrl: String?, autoReturn: PXAutoReturn?) {
         self.points = points
         self.discounts = discounts
         self.crossSelling = crossSelling
@@ -27,6 +32,11 @@ struct PXPointsAndDiscounts: Decodable {
         self.customOrder = customOrder
         self.expenseSplit = expenseSplit
         self.paymentMethodsImages = paymentMethodsImages
+        self.primaryButton = primaryButton
+        self.secondaryButton = secondaryButton
+        self.redirectUrl = redirectUrl
+        self.backUrl = backUrl
+        self.autoReturn = autoReturn
     }
 
     enum PointsAndDiscountsCodingKeys: String, CodingKey {
@@ -38,6 +48,11 @@ struct PXPointsAndDiscounts: Decodable {
         case customOrder = "custom_order"
         case expenseSplit = "expense_split"
         case paymentMethodsImages = "payment_methods_images"
+        case primaryButton = "primary_button"
+        case secondaryButton = "secondary_button"
+        case redirectUrl = "redirect_url"
+        case backUrl = "back_url"
+        case autoReturn = "auto_return"
     }
 
     init(from decoder: Decoder) throws {
@@ -50,6 +65,11 @@ struct PXPointsAndDiscounts: Decodable {
         let customOrder: Bool? = try container.decodeIfPresent(Bool.self, forKey: .customOrder)
         let expenseSplit: PXExpenseSplit? = try container.decodeIfPresent(PXExpenseSplit.self, forKey: .expenseSplit)
         let paymentMethodsImages: [String: String]? = try container.decodeIfPresent([String: String].self, forKey: .paymentMethodsImages)
-        self.init(points: points, discounts: discounts, crossSelling: crossSelling, viewReceiptAction: viewReceiptAction, topTextBox: topTextBox, customOrder: customOrder, expenseSplit: expenseSplit, paymentMethodsImages: paymentMethodsImages)
+        let primaryButton: PXButton? = try container.decodeIfPresent(PXButton.self, forKey: .primaryButton)
+        let secondaryButton: PXButton? = try container.decodeIfPresent(PXButton.self, forKey: .secondaryButton)
+        let redirectUrl: String? = try container.decodeIfPresent(String.self, forKey: .redirectUrl)
+        let backUrl: String? = try container.decodeIfPresent(String.self, forKey: .backUrl)
+        let autoReturn: PXAutoReturn? = try container.decodeIfPresent(PXAutoReturn.self, forKey: .autoReturn)
+        self.init(points: points, discounts: discounts, crossSelling: crossSelling, viewReceiptAction: viewReceiptAction, topTextBox: topTextBox, customOrder: customOrder, expenseSplit: expenseSplit, paymentMethodsImages: paymentMethodsImages, primaryButton: primaryButton, secondaryButton: secondaryButton, redirectUrl: redirectUrl, backUrl: backUrl, autoReturn: autoReturn)
     }
 }