+++ MercadoPagoSDK/MercadoPagoSDK/UI/PXFooter/PXFooterProps.swift
@@ -16,8 +16,9 @@ final class PXFooterProps: NSObject {
     weak var animationDelegate: PXAnimatedButtonDelegate?
     var pinLastSubviewToBottom: Bool
     var termsInfo: PXTermsDto?
+    var andesButtonConfig: PXAndesButtonConfig
 
-    init(buttonAction: PXAction? = nil, linkAction: PXAction? = nil, useAndesButtonForLinkAction: Bool = false, primaryColor: UIColor? = ThemeManager.shared.getAccentColor(), animationDelegate: PXAnimatedButtonDelegate? = nil, pinLastSubviewToBottom: Bool = true, termsInfo: PXTermsDto? = nil) {
+    init(buttonAction: PXAction? = nil, linkAction: PXAction? = nil, useAndesButtonForLinkAction: Bool = false, primaryColor: UIColor? = ThemeManager.shared.getAccentColor(), animationDelegate: PXAnimatedButtonDelegate? = nil, pinLastSubviewToBottom: Bool = true, termsInfo: PXTermsDto? = nil, andesButtonConfig: PXAndesButtonConfig = PXAndesButtonConfig()) {
         self.buttonAction = buttonAction
         self.linkAction = linkAction
         self.useAndesButtonForLinkAction = useAndesButtonForLinkAction
@@ -25,5 +26,6 @@ final class PXFooterProps: NSObject {
         self.animationDelegate = animationDelegate
         self.pinLastSubviewToBottom = pinLastSubviewToBottom
         self.termsInfo = termsInfo
+        self.andesButtonConfig = andesButtonConfig
     }
 }